Fix data type

+++ feedback.go
@@ -13,8 +13,8 @@ var (
 	queryGetFeedbackCountGuild string
 )
 
-func (c *Client) GetAverageFeedbackRatingGuild(ctx context.Context, guildId uint64) (float32, error) {
-	var rating float32
+func (c *Client) GetAverageFeedbackRatingGuild(ctx context.Context, guildId uint64) (float64, error) {
+	var rating float64
 	if err := c.client.QueryRow(ctx, queryGetAverageFeedbackRatingGuild, guildId).Scan(&rating); err != nil {
 		return 0, err
 	}
